Remove ShardClusterSnapshotRestoreIT and fold test in DataStreamsSnapshotsIT
